Add unit tests and separate integration tests
