+++ orkes-redis-queues/build.gradle
@@ -5,12 +5,48 @@ dependencies {
     //Guava
     implementation "com.google.guava:guava:${versions.revGuava}"
 
-    testImplementation "junit:junit:4.11"
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
+    testImplementation 'org.mockito:mockito-junit-jupiter:5.7.0'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
-    testImplementation "org.testcontainers:testcontainers:1.17.2"
 }
 
 test {
     useJUnitPlatform()
 }
+
+// integration test configurations
+sourceSets {
+    intTest {
+        java {
+            compileClasspath += main.output + test.output
+            runtimeClasspath += main.output + test.output
+            srcDir file('src/integration-test/java')
+        }
+    }
+}
+
+configurations {
+    intTestImplementation.extendsFrom testImplementation
+    intTestRuntimeOnly.extendsFrom testRuntimeOnly
+}
+
+dependencies {
+    intTestImplementation "org.testcontainers:testcontainers:1.17.2"
+}
+
+tasks.register('integrationTest', Test) {
+    description = 'Runs integration tests against real persistent stores.'
+    group = 'verification'
+
+    testClassesDirs = sourceSets.intTest.output.classesDirs
+    classpath = sourceSets.intTest.runtimeClasspath
+    shouldRunAfter test
+
+    useJUnitPlatform()
+
+    testLogging {
+        events "passed"
+    }
+}
+
+check.dependsOn integrationTest
